[exporter/opensearch] Support multiple encoding schemas for logs
